move infer() to inference.py
